requirements.txt
PyQt6==6.7.0
cryptography==42.0.5
pycryptodome==3.20.0
pyjwt==2.8.0
matplotlib==3.8.4
qrcode[pil]